requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
psycopg2-binary==2.9.9
click==8.3.1
colorama==0.4.6
fastapi==0.128.0
h11==0.16.0
idna==3.11
beautifulsoup4==4.12.3
openpyxl==3.1.5
pandas==2.3.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
